+++ notification/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class NotificationsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'notifications'
+    name = 'notification'
