+++ user/urls.py
@@ -5,9 +5,9 @@ from rest_framework_simplejwt.views import (
     TokenVerifyView,
 )
 
-from users.views import CreateUserView, ManageUserView
+from user.views import CreateUserView, ManageUserView
 
-app_name = "users"
+app_name = "user"
 
 urlpatterns = [
     path("", CreateUserView.as_view(), name="user-register"),
